API/requirements.txt
pip~=25.1.1
flask-marshmallow~=0.15
marshmallow~=3.20.0
flask-sqlalchemy~=3.1.1
werkzeug~=2.3.7
flask~=2.3.3
python-dotenv~=1.0.0
flask-Cors~=4.0.0
flask_jwt_extended
Flask-JWT-Extended~=4.5.3
mysqlclient~=2.2.0
flask-Bcrypt~=1.0.1
pytest~=7.4.3
python-dotenv~=1.0.0
flask-Migrate~=4.0.5
alembic~=1.12.1
SQLAlchemy~=2.0.23
click~=8.1.7
marshmallow-sqlalchemy~=0.28.0
